backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# LangChain + LangGraph stack (0.3.x family, kept compatible)
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.6
langgraph==0.2.45
langchain-text-splitters==0.3.2

faiss-cpu==1.9.0.post1
pillow==11.0.0
pytesseract==0.3.13